requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
colour==0.1.5
numpy==1.18.0
pillow==6.2.1
progressbar2==3.47.0
python-utils==2.3.0       # via progressbar2
six==1.13.0               # via progressbar2, python-utils
